Added new directory for Window Utils. Renamed Notification to notification_helper and moved it to helpers folder.
